requirements.txt

torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
numpy>=1.26.0
pandas>=2.2.0
Pillow>=10.0.0
tqdm>=4.66.0
scikit-learn>=1.4.0
streamlit>=1.36.0
flask>=3.0.0
click>=8.1.7
